the Barrel calorimeter has now an 8mm aluminum plate at the inside of the Barrel. This volume is now also visible by GEANT. the same is done for the 3x3 readout version.
The CDC are now defined according to the Gluex doc 979 with the correct number 
of straws per layer and the correct radii. Also the axial layers are shifted/rotated
w.r.t. eachother layer so avoid lineups.

all this changes are reflected in the fortran geometry file need for the MC and the
file is updated in the appropirate directory HDGeant.